+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2022-06-22
+
 - Add DNS record for bastion nodes.
 
 ## [0.4.0] - 2022-06-02
@@ -25,7 +27,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] - 2022-05-05
 
-[Unreleased]: https://github.com/giantswarm/dns-operator-gcp/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/giantswarm/dns-operator-gcp/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/giantswarm/dns-operator-gcp/compare/v0.4.0...v0.3.0
 [0.4.0]: https://github.com/giantswarm/dns-operator-gcp/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/giantswarm/dns-operator-gcp/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/giantswarm/dns-operator-gcp/compare/v0.1.0...v0.2.0